@@ -29,9 +29,9 @@ packages = find:
 install_requires =
     matplotlib
     napari
-    numpy
+    numpy>=1.23
     tinycss2
-python_requires = >=3.9
+python_requires = >=3.10
 include_package_data = True
 package_dir =
     =src
@@ -47,11 +47,10 @@ napari.manifest =
 
 [options.extras_require]
 docs =
-    napari[all]==0.4.19rc3
+    napari[all]
     numpydoc
     pydantic<2
     pydata-sphinx-theme
-    qtgallery
     sphinx
     sphinx-automodapi
     sphinx-gallery
