@@ -13,7 +13,7 @@
 # import os
 # import sys
 # sys.path.insert(0, os.path.abspath('.'))
-import qtgallery
+from sphinx_gallery import scrapers
 
 # -- Project information -----------------------------------------------------
 
@@ -35,18 +35,58 @@ extensions = [
     "sphinx.ext.intersphinx",
 ]
 
+
+def reset_napari(gallery_conf, fname):  # type: ignore[no-untyped-def]
+    from napari.settings import get_settings
+    from qtpy.QtWidgets import QApplication
+
+    settings = get_settings()
+    settings.appearance.theme = "dark"
+
+    # Disabling `QApplication.exec_` means example scripts can call `exec_`
+    # (scripts work when run normally) without blocking example execution by
+    # sphinx-gallery. (from qtgallery)
+    QApplication.exec_ = lambda _: None
+
+
+def napari_scraper(block, block_vars, gallery_conf):  # type: ignore[no-untyped-def]
+    """Basic napari window scraper.
+
+    Looks for any QtMainWindow instances and takes a screenshot of them.
+
+    `app.processEvents()` allows Qt events to propagateo and prevents hanging.
+    """
+    import napari
+
+    imgpath_iter = block_vars["image_path_iterator"]
+
+    if app := napari.qt.get_app():
+        app.processEvents()
+    else:
+        return ""
+
+    img_paths = []
+    for win, img_path in zip(
+        reversed(napari._qt.qt_main_window._QtMainWindow._instances),
+        imgpath_iter,
+        strict=False,
+    ):
+        img_paths.append(img_path)
+        win._window.screenshot(img_path, canvas_only=False)
+
+    napari.Viewer.close_all()
+    app.processEvents()
+
+    return scrapers.figure_rst(img_paths, gallery_conf["src_dir"])
+
+
 sphinx_gallery_conf = {
     "filename_pattern": ".",
-    "image_scrapers": (qtgallery.qtscraper,),
-    "reset_modules": (qtgallery.reset_qapp,),
+    "image_scrapers": (napari_scraper,),
+    "reset_modules": (reset_napari,),
 }
+suppress_warnings = ["config.cache"]
 
-qtgallery_conf = {
-    "xvfb_size": (640, 480),
-    "xvfb_color_depth": 24,
-    "xfvb_use_xauth": False,
-    "xfvb_extra_args": [],
-}
 
 numpydoc_show_class_members = False
 automodapi_inheritance_diagram = True
