@@ -26,7 +26,7 @@ def test_interval():
     assert 10 not in interval
 
     with pytest.raises(ValueError, match="must be an integer"):
-        "string" in interval  # type: ignore
+        assert "string" in interval  # type: ignore[operator]
 
     with pytest.raises(ValueError, match="must be <= upper_bound"):
         Interval(5, 3)
